Rename versions to common, split source code into multiple files
Additionally added dynamic creation of mod files
