+++ releaser/go.mod
@@ -3,6 +3,7 @@ module go.opentelemetry.io/build-tools/releaser
 go 1.15
 
 require (
+	github.com/go-git/go-git/v5 v5.4.2
 	github.com/spf13/cobra v1.2.1
 	github.com/spf13/viper v1.8.1
 	github.com/stretchr/testify v1.7.0
