+++ releaser/internal/common/test_data/new_module_versioning/versions_valid.yaml
