--- releaser/internal/verify/test_data/new_verification/go.mod
@@ -1,9 +1,0 @@
-module go.opentelemetry.io/build-tools/releaser/internal/verify/testroot
-
-go 1.16
-
-require (
-	go.opentelemetry.io/build-tools/releaser/internal/verify/test/test1 v1.2.3-RC1+meta
-	go.opentelemetry.io/build-tools/releaser/internal/verify/test/test2 v1.2.3-RC1+meta
-	go.opentelemetry.io/build-tools/releaser/internal/verify/test3 v0.1.0
-)
