--- releaser/internal/versions/test_data/build_module_path_map/go.mod
@@ -1,3 +1,0 @@
-module go.opentelemetry.io/testroot
-
-go 1.16
