--- releaser/internal/versions/test_data/build_module_path_map/test/go.mod
@@ -1,3 +1,0 @@
-module go.opentelemetry.io/test3
-
-go 1.16
