--- releaser/internal/versions/test_data/new_module_versioning/go.mod
@@ -1,3 +1,0 @@
-module go.opentelemetry.io/testroot
-
-go 1.16
