--- releaser/internal/versions/test_data/new_module_versioning/test/go.mod
@@ -1,3 +1,0 @@
-module go.opentelemetry.io/test3
-
-go 1.16
