--- releaser/internal/versions/test_data/new_module_versioning/test/test1/go.mod
@@ -1,3 +1,0 @@
-module "go.opentelemetry.io/test/test1"
-
-go 1.16
